+++ src/components/layout/Header/DeleteandLog.jsx
@@ -26,12 +26,21 @@ const TextButton = styled.button`
 
 const DeleteandLog = ({ onToast, onPost }) => {
   const navigate = useNavigate();
+
   return (
     <Container>
       <TextButton
         style={{ color: '#FF3F3F' }}
         onClick={() => {
-          navigate('/');
+          navigate(`/`, {
+            state: {
+              toastData: {
+                show: true,
+                type: 'positive',
+                message: '삭제되었습니다!',
+              },
+            },
+          });
         }}
       >
         삭제하기
